requirements.txt
# --- Core Framework ---
llama-index-core==0.14.12
llama-index-readers-file==0.5.6

# --- AI Models (Gemini & Hugging Face) ---
llama-index-llms-gemini==0.6.1
llama-index-embeddings-huggingface==0.6.1
google-generativeai==0.8.6

# --- Vector Database (Persistence) ---
chromadb==1.4.0
llama-index-vector-stores-chroma==0.5.5

# --- Data Parsing ---
docx2txt==0.9
pypdf==6.5.0
pandas==2.2.3
openpyxl==3.1.5

# --- UI & Environment ---
streamlit==1.52.2
python-dotenv==1.2.1